move duration to utilities
use for dataset lastAccessed
move tests from phantomjs to karma

refs #1432
