+++ modules/Hoot/tools/utilities.js
@@ -5,7 +5,7 @@
  *******************************************************************************************************/
 
 import _forEach from 'lodash-es/forEach';
-
+import dayjs from 'dayjs';
 import { t } from '../../util/locale';
 
 export const getBrowserInfo = () => {
@@ -131,4 +131,41 @@ export const formatBbox = str => {
     }
 
     return `${minx},${miny},${maxx},${maxy}`;
+};
+
+export const duration = (start, end, started) => {
+    let duration,
+        diff = dayjs(end).diff(dayjs(start), 'seconds');
+
+    function calcDiff(diff, unit) {
+        diff = Math.floor(diff);
+        let calc;
+        if (diff < 1) {
+            calc = `less than a ${unit}`;
+        } else if (diff === 1) {
+            let article = unit === 'hour' ? 'an' : 'a';
+            calc = `${article} ${unit}`;
+        } else if (diff < 5) {
+            calc = `a few ${unit}s`;
+        } else {
+            calc = `${diff} ${unit}s`;
+        }
+        return calc;
+    }
+
+    if (diff < 60) {
+        duration = calcDiff(diff, 'second');
+    } else if ((diff / 60) < 60) {
+        duration = calcDiff(Math.floor(diff / 60), 'minute');
+    } else if ((diff / 3600) < 24) {
+        duration = calcDiff(Math.floor(diff / 3600), 'hour');
+    } else {
+        duration = calcDiff(Math.floor(diff / 86400), 'day');
+    }
+
+    if (started) {
+        duration += ' ago';
+    }
+
+    return duration;
 };
