+++ modules/Hoot/ui/managePanel/jobs.js
@@ -1,8 +1,8 @@
 import Tab            from './tab';
-import dayjs from 'dayjs';
 import ProgressBar    from 'progressbar.js';
 import DifferentialStats from '../modals/differentialStats';
 import JobCommandInfo from '../modals/jobCommandInfo';
+import { duration } from '../../tools/utilities';
 
 const getJobTypeIcon = Symbol('getJobTypeIcon');
 
@@ -18,43 +18,6 @@ export default class Jobs extends Tab {
 
         this.name = 'Jobs';
         this.id   = 'util-jobs';
-    }
-
-    duration(start, end, started) {
-        let duration,
-            diff = dayjs(end).diff(dayjs(start), 'seconds');
-
-        function calcDiff(diff, unit) {
-            diff = Math.floor(diff);
-            let calc;
-            if (diff < 1) {
-                calc = `less than a ${unit}`;
-            } else if (diff === 1) {
-                let article = unit === 'hour' ? 'an' : 'a';
-                calc = `${article} ${unit}`;
-            } else if (diff < 5) {
-                calc = `a few ${unit}s`;
-            } else {
-                calc = `${diff} ${unit}s`;
-            }
-            return calc;
-        }
-
-        if (diff < 60) {
-            duration = calcDiff(diff, 'second');
-        } else if ((diff / 60) < 60) {
-            duration = calcDiff(Math.floor(diff / 60), 'minute');
-        } else if ((diff / 3600) < 24) {
-            duration = calcDiff(Math.floor(diff / 3600), 'hour');
-        } else {
-            duration = calcDiff(Math.floor(diff / 86400), 'day');
-        }
-
-        if (started) {
-            duration += ' ago';
-        }
-
-        return duration;
     }
 
     render() {
@@ -207,7 +170,7 @@ export default class Jobs extends Tab {
 
                 //Start
                 props.push({
-                    span: [{ text: this.duration(d.start, Date.now(), true) }]
+                    span: [{ text: duration(d.start, Date.now(), true) }]
                 });
 
                 //Progress bar
@@ -426,12 +389,12 @@ export default class Jobs extends Tab {
 
                 //Start
                 props.push({
-                    span: [{text: this.duration(d.start, Date.now(), true) }]
+                    span: [{text: duration(d.start, Date.now(), true) }]
                 });
 
                 //Duration
                 props.push({
-                    span: [{text: this.duration(d.start, d.end) }]
+                    span: [{text: duration(d.start, d.end) }]
                 });
 
                 //Actions
