+++ test/hoot/spec/layerManager.js
